requirements.txt
fastapi
uvicorn[standard]
httpx
openai
aiosqlite
pydantic[email]
python-dotenv
